Add Encode and EncodeInPython2 classes
Encode allows one to encode Unicode strings to byte strings. The
EncodeInPython2 class is useful in cases where the output stream
is already expecting Unicode strings (such as the default stdout/stderr
IOTextBase classes being used in python3 when attached to a terminal).

In python2 it behaves just as Encode() (since stdout/stderr is a dumb
file there and we really want to explicitly convert all Unicode
strings to byte strings before calling write())

In python3 it will be a no-op, hopefully forwarding each line to
Unicode-aware IOTextBase instance.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>