Convert extcmd to a package
Tests had to be moved as they would no longer be able to find the README
file after the extcmd_test.py file was moved to extcmd/tests.py.

Apparently this is caused by setuptools not copying stuff from MANIFEST to
the build directory where tests were being executed. This also makes
it possible to test an installed version of extcmd.

Tests had to be minimally modified, all occurences of \n had to be escaped
to \\n, otherwise \n got parsed at the python readed level and the
meaning was changed

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>